move materials into their own domain module
