+++ src/application/tonnage-monitoring/integration.test.js
@@ -4,10 +4,7 @@ import { it, DATABASE_NAME } from '#vite/fixtures/mongo-client.js'
 import { aggregateTonnageByMaterial } from './aggregate-tonnage.js'
 import { PROCESSING_TYPES } from '#domain/summary-logs/meta-fields.js'
 import { WASTE_RECORD_TYPE } from '#domain/waste-records/model.js'
-import {
-  MATERIAL,
-  GLASS_RECYCLING_PROCESS
-} from '#domain/organisations/model.js'
+import { MATERIAL, GLASS_RECYCLING_PROCESS } from '#domain/materials.js'
 import { SUMMARY_LOG_ROW_STATES_COLLECTION_NAME } from '#waste-records/repository/mongodb.js'
 import {
   WASTE_BALANCE_EVENTS_COLLECTION_NAME,
