+++ src/forms-submission-data/registration/transform-registration.test.js
@@ -5,9 +5,8 @@ import { validateRegistration } from '#repositories/organisations/schema/validat
 import exporter from '#data/fixtures/ea/registration/exporter.json' with { type: 'json' }
 import reprocessorAllMaterials from '#data/fixtures/ea/registration/reprocessor-all-materials.json' with { type: 'json' }
 import reprocessorSepa from '#data/fixtures/ea/registration/reprocessor-all-materials-sepa-handcrafted.json' with { type: 'json' }
+import { GLASS_RECYCLING_PROCESS, MATERIAL } from '#domain/materials.js'
 import {
-  GLASS_RECYCLING_PROCESS,
-  MATERIAL,
   REGULATOR,
   TIME_SCALE,
   VALUE_TYPE,
