+++ src/packaging-recycling-notes/repository/schema.js
@@ -1,9 +1,6 @@
 import Joi from 'joi'
 
-import {
-  MATERIAL,
-  GLASS_RECYCLING_PROCESS
-} from '#domain/organisations/model.js'
+import { MATERIAL, GLASS_RECYCLING_PROCESS } from '#domain/materials.js'
 import { PRN_STATUS } from '#packaging-recycling-notes/domain/model.js'
 
 const materialValues = Object.values(MATERIAL)
