+++ src/packaging-recycling-notes/routes/external-transition-handler.test.js
@@ -2,7 +2,8 @@ import { StatusCodes } from 'http-status-codes'
 import { randomUUID } from 'node:crypto'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
-import { MATERIAL, REGULATOR } from '#domain/organisations/model.js'
+import { MATERIAL } from '#domain/materials.js'
+import { REGULATOR } from '#domain/organisations/model.js'
 import {
   PRN_ACTOR,
   PRN_STATUS,
