+++ src/packaging-recycling-notes/routes/reject.test.js
@@ -3,7 +3,8 @@ import { randomUUID } from 'node:crypto'
 import { ObjectId } from 'mongodb'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
-import { MATERIAL, REGULATOR } from '#domain/organisations/model.js'
+import { MATERIAL } from '#domain/materials.js'
+import { REGULATOR } from '#domain/organisations/model.js'
 import { createInMemoryFeatureFlags } from '#feature-flags/feature-flags.inmemory.js'
 import { PRN_STATUS } from '#packaging-recycling-notes/domain/model.js'
 import { createInMemoryPackagingRecyclingNotesRepository } from '#packaging-recycling-notes/repository/inmemory.plugin.js'
