+++ src/reports/repository/contract/findReportById.contract.js
@@ -1,7 +1,8 @@
 import { describe, beforeEach, expect } from 'vitest'
 import { randomUUID } from 'node:crypto'
 import { REPORT_STATUS } from '#reports/domain/report-status.js'
-import { MATERIAL, WASTE_PROCESSING_TYPE } from '#domain/organisations/model.js'
+import { MATERIAL } from '#domain/materials.js'
+import { WASTE_PROCESSING_TYPE } from '#domain/organisations/model.js'
 import {
   buildCreateReportParams,
   DEFAULT_ORG_ID,
