+++ src/repositories/organisations/contract/accreditation-link.contract.js
@@ -1,6 +1,5 @@
+import { GLASS_RECYCLING_PROCESS, MATERIAL } from '#domain/materials.js'
 import {
-  GLASS_RECYCLING_PROCESS,
-  MATERIAL,
   REGISTRATION_STATUS,
   REPROCESSING_TYPE,
   WASTE_PROCESSING_TYPE
