+++ src/routes/v1/tonnage-monitoring/response.schema.js
@@ -1,8 +1,6 @@
 import Joi from 'joi'
-import {
-  TONNAGE_MONITORING_MATERIALS,
-  WASTE_PROCESSING_TYPE
-} from '#domain/organisations/model.js'
+import { REPORTING_MATERIALS } from '#domain/materials.js'
+import { WASTE_PROCESSING_TYPE } from '#domain/organisations/model.js'
 import { capitalize } from '#common/helpers/formatters.js'
 
 const WASTE_PROCESSING_TYPES = Object.values(WASTE_PROCESSING_TYPE).map(
@@ -33,7 +31,7 @@ const START_YEAR = 2026
 const MAX_YEAR = 2100
 const materialTonnageSchema = Joi.object({
   material: Joi.string()
-    .valid(...TONNAGE_MONITORING_MATERIALS)
+    .valid(...REPORTING_MATERIALS)
     .required(),
   year: Joi.number().integer().min(START_YEAR).max(MAX_YEAR).required(),
